requirements.txt
binance-connector==3.12.0
pandas==2.3.1
numpy==2.2.6
python-dotenv==1.0.1
